requirements.txt
streamlit
tensorflow==2.18.0
numpy
pandas
plotly
Pillow
wikipedia
streamlit_lottie